+++ uefi/CHANGELOG.md
@@ -8,6 +8,7 @@
 - Added `proto::hii::config::HiiConfigAccess`.
 - Added `proto::hii::config_str::ConfigurationString`.
 - Added `proto::acpi::AcpiTable`.
+- Added `proto::hii::database::HiiDatabase`.
 
 ## Changed
 - **Breaking:** `boot::stall` now take `core::time::Duration` instead of `usize`.